Kraken : corrections to the connections patch
Several corrections to the patch that add connections to the starting point.

Since we want to avoid stuff like walking + connection + TC, we explore the different connections of the possible starting points to add them to the possible starting points

There were several problems but mainly the different graphs (walking, bike, car) were not taken correctly into account.

We now project the starting and ending point to the graph corresponding to the transport mode.