Kraken: changes in graph construction
changes in graph construction for bike sharing system.

The main idea is for each bike sharing station to find the
nearest edge on the walking graph and the nearest edge on the
biking graph.

we have to add a new edge between those 2 with the walking
time between the edges + the time to take of to hang back
the bike