Fare: SectionKey constructor change
To minimize impact of code, for the moment we try not the change the
algorithm structures, so we only change the way SectionKeys are created.

unfortunatly to do that some refactory to routing::PathItem had to be
done.

Add boost serialization to

For the moment Unit tests are broken :/ so they are commented