Jormungandr: fix the cache, all object needs to be merge in the session
When we retrieve an object from the cache, it need to be merge with
SQLAlchemy, else all objectProxy would throw a exception when used.
The parameter "load" of the merge method is set to false for avoid a
request to the database(this will negate the benefice of the cache)

And for better performance, all relation to instances are set to
"lazy=joined", with that all instances are loaded without additional
queries, and allow instances to be cached too. Without this all call
would need to fetch the instances directly from the database.