TimeTable: change exceptions
the exceptions on the calendar association were messed up, they were
opposite to the one we should put.

In the meantime, the period ends have been changed to add one day to
them since the GTFS/Fusio end dates have to be in the period (and they
were not since boost period work as a classic range so the end is the
first date after the period).