Kraken: fix production_period in navitia
On date periods end() return one day past the last day in the period so
all call to end() has to be replaced by last().

Remove a test on timetables who check if the date and the duration was
on the production_period, now we only check the date. So even if the
requested period for the timetable overflow the production_period the
request will be served but will possibly return no circulation.

http://jira.canaltp.fr/browse/NAVITIAII-964