journeys: add the posibility to have different fallback duration for each mode
Split the "max_duration_to_pt" for each fallback mode. We keep the
"max_duration_to_pt" in the interface for not breaking the compatibily.

The the max duration for car is now set to 30 min.

The "max_duration_to_pt" field in the protobuf is keeped for compatibily
during the migration to the next version. It will be removed in another
pull request.