Add notes on lines and routes for journeys and schedules API
A field "notes" has been added to the PtDisplayInfo protobuf message, it
contain notes for the route, the line and the vehicle journey.

In the journey API the notes are referenced in the item "links" of a
section.

In the route schedules API the notes are in ```route_schedules.table.headers.links```
and in ```route_schedules.links```. Notes on lines and routes are duplicated, one time on the global links
and on each headers.

And finally for Stop Schedules the notes are in the links of each items: ```stop_chedules[].links```