Kraken: fix a bug on car or vls at the arrival, we were updated the last section
We add some coherence between the pt part of the journey and the street
network part, we assure that the departure of the streetnetwork section
is the arrival stop_point of the last pt section.
We were always updated the last section added but a with car or vls
there is more than one section add.