Kraken: remove duplicate impact on vehicle journey
For blocking impact we have to update the vehicle_journeys attached at
the impacted object, typically a line or a route. For being able to
update a impact we have to store each impact impacting the vj. But this
impacts must not be displayed for this vj else we will have duplicate
items on the response.

So impacts not directly on a vj are stored in a new attributes:
```impacted_by```.