jormungandr: filter on too long duration can select standart by duration
it's now possible to choose the criteria used for select the journey
used a standard for calculate the max_duration, before it was only on
departure_time/arrival_time, the select on duration has been added.

Before journey with other fallback mode than walking were ignored, we
can now choose the modes allowed.