Kraken: don't mess with the origin of the streetnetwork section
When we create the sections for the street network (eventually more than
one) we set the departure and the arrivals of these to the requested
departure and to the stop_point used for the PT part, this way we have
some coherence in the response.

Before this change only the last section of the streetnetwork was
modified, so the origin of the last section (typically the bss_put_back
or park section) was the origin of journey.

Refer to: http://jira.canaltp.fr/browse/NAVITIAII-1663