ed2nav: fix bug with lacking edges on street network
When choosing nodes to be ignored in the graph we were using the
vertex_t of the node and not it's id, so we were deleting the wrong
node, that caused some edge to be deleted too.

Ironically when removing small connected component we were in fact
randomly created some of them.