ed: fix a bug when a way overlap 2 admins
When a way was touching 2 admins, the admin was propagated to the other ways.
We fix that. For example, if we have way1 at Bordeaux, way2 at Bordeaux and Bègles
and way3 at Bègles, before this fix, way1, way2 and way3 were at Bordeaux. After,
way1 is at Bordeaux, way2 at Bordeaux and Bègles, and way3 at Bègles.