kraken: don't listen to request before data are loaded
We also used a named queue for rabbitmq, this way if there is a small
disconnection we don't loose messages. The drawbacks are that we can't
anymore start two kraken for the same instance on the same host. And if
an host/instance is deleted, the queues associated won't be delete from
rabbitmq.