jormungandr: fix a bug in stat manager when the first init failed
If we weren't able to connect to rabbitmq at the start of jormungandr
the producer is not created, and since we use is to call publish an
exception is raise since producer is None. We were only catching
exception directly link to the protocols, so we didn't try to reconnect.