jormungandr: fix a bug depending of the order of the parameter
Typically on admin to admin journeys we can get a journey like this:
Crowfly > PT > Crowfly
The mode of the crowfly is the one asked to kraken, so kraken will give
the exact same response for all fallback mode except for the mode used
(but we don't care since it's a teleportation)
And since all journeys are the same, jormungandr only keep the first
one, eventually the one with car fallback. Not cool but why not, except
that we only want one car journey so we remove a few journeys.
Paradoxically enabling car fallback was reducing the number of solution
found.

With this patch fallback modes are used in a logical order:
Walking > bss > bike > car

Refer to: http://jira.canaltp.fr/browse/NAVITIAII-1909