Kraken can now compute direct path on car
In kraken, when we compute a direct path for car, we want an arrival on
the walking graph, so we will park the car at a parking and finish at
foot.

Jormungandr has the responsability to choose if a direct path has to be
compute. For the default scenario, direct path are always compute except
for car and for journey of admin to admin type.
New default always ask for a direct path.