jormun: filter short fallback on car and bike in new_default
min_car and min_bike are back in business! We don't want to display very
short fallback in car or bike.
Almost nobody take it's car for 2 minutes.
This is a temporary solution we will hopefully find a better rule for
removing ugly journey.
Prior to this PR min_car and min_bike will be set to 0 on all instance
using new_default except fr-auv