kraken: revert new output of perturbation on line_section
We were using impacted_stops to output impacted stop points on line_section but
it's designed to output stop_times and not stop_points.
To prevent any confusion we are keeping the current way to output impact on line_section.
A proper way to display them will be done in the future.