osm2ed: fix a bug with private way part of an administrative boundary
We didn't import private way even when they were part of a boundary, we
now keep them.

There is room for improvement, a lot of useless way seems to be
imported, all ways part of a boundary are imported, they aren't use in
the streetnetwork but we insert them.
Since this component should progressively disappear and is not covered
by any test I think the best call is to let it this way.