jormun: add the ability to use transient zmq socket
By default jormungandr keep it's zmq socket between calls, this PR give
us the ability to use transient socket: each time a call to kraken has
to be made we create a new connection. The main advantage is that the
load balancing will be done correctly between the available krakens, of
course this has consequences on performance, it cost us a tcp roundtrip,
so if the network is already slow thing are getting worse. When using
IPC it's free: I wasn't able to see any difference between bench.

This PR will allow us to do some tests on production with this setup, it
will let us measure the impacts on response time and load on our network's
equipments.

A few informations on load balancing: We actually load balance so that
each kraken has the same number of connection established, so in
production that at least 60 connections to distribute over our 4 kraken.
When we do parallel call to a kraken new connections are created.
A parallel journey can use 10 connections, and their is absolutely no
guarantee that this connections will be on different kraken, so it's
possible to run all of our request on one kraken and queuing some of the
request because no thread are available on this kraken, but the others
can have nothing to do at this time.