jormun: removal of thread ping
Currently each jormun call every instance each 10 seconds to check if
it's still alive. It's useless since we don't do anything of this
information and it generate a lot unneeded request for kraken (remember
their is 60 jormungandr in production)

I kept the thread just to ensure that we call each instance at least one
time after the startup, this way we have the geom and the timezone of
the instance.

With this change after the startup jormungandr will only call kraken
when handling a request this will allow us to spawn kraken only when
we need (something based on xinetd probably) it on some testing environments.