Kraken: configuration from environment variables
This commit introduce the configuration of kraken from the environment
variables and of course make the configuration file optional.

These variable are very badly name since we want to be able to read all
configuration files and a new option_descriptions or a mapping table
between environment variable and file's key is error prone since every
change has to be made two time at least.

So we have ugly names with the following form: `KRAKEN_{SECTION}_{FIELD_NAME}`
This will converted to `{SECTION}.{FIELD_NAME}`.

So to set the zmq_socket we need to define a variable named
`KRAKEN_GENERAL_zmq_socket`

Since the configuration file was also used for logger configuration, I
added two new properties so we can configure the logging with
environment variables too.