ed: import boarding and alighting duration from NTFS
We have two new optional fields in NTFS:
  - boarding_duration,
  - alighting_duration.
Representing the necessary time taken by the boarding process or
alighting process. We store the boarding and alighting time into
database. The boarding_time is the departure_time minus the boarding
duration and the alighting is the arrival_time plus the alighting
duration. These values can't be negative.