all: use boarding and alighting time in place of departure arrival
Replace departure_time / arrival_time in the routing by boarding and
alighting_time.
The main issue is that we can't use datetime::shift anymore because it
was supposing that the departure_time were ordered chronologicaly when
moving to the next stop_time, otherwise, it was a day shift. It's not
true anymore since the boarding duration at a stop could vary. Therefore
we maintain the shift time directly in raptor and only work with non
shifted times.