Jormungandr: fix some links
in some wizard decorator adding link we were using the json response
content to deduce the type of the returned object.

since 'disruptions' is a list we can get in all navitia's responses (and
in /disruptions responses) we consider the type is a disruption only if
we do not find another list (like stop_areas, networks, ...)

this was creating arbitrary behavior (and leading serpy tests to be
different)

as it is a random behavior I don't see an easy way to test this.

Note: beware, this will break artemis