time_tables: add failing passages test
There is a bug when getting previous departures / arrivals on the
first production day where the code return infinite next departures.
We also add a test for next departures on the last production day,
since we now want to just limit the requested duration, as long as the
request datetime is in the production period, instead of throwing
an out of bounds error.