Jormun & Tyr: add a default coord to a user
we add the possibility to set a default coordinate to a user.

This coordinate is given to the autocomplete service (bragi) if no
coordinate is set in the query (the coord is used to sort the result by
proximity to this coord)

Note: for the moment it is not possible, for a user with a coord to
'unset' the coord in the `/places` api (he can give another coord, but
cannot call with no coord).

I don't know if it's important