Kraken: don't segfault when 2 different disruptions modify the same vj
If we have a disruption updating the stop times of a vj, and then
another disruption on the same vj with another id, we have dangling
pointers.  It should not arrive but better to secure that.

This PR fixes that in a simple way:
  * the disruption is kept, as some tests require it.
  * the disruption will be printed when it applies, even if the other
    disruption override it.
  * the impacted stops will be printed relative to UTC instead of
    local.

This is safe, but not the prettiest. As it should not happen, keep it simple.