Jormungandr: get all bss stands in one call to jcdecaux api
With this PR we only call the jcdecaux's api one time for all the
stations, this result is then cached so it be reuse for other request.
This will greatly improve performance on call like this one:
https://api.navitia.io/v1/coverage/fr-idf/poi_types/poi_type%3Aamenity%3Abicycle_rental/pois?count=500
In most case the hard timeout of 30sec was hit before navitia can
respond.