Jormungandr: disabling of sql connection pooling
This PR give us the possibility to disable connection pooling to the
postgresql database.
Actually each jormungrandr process keep it's sql connection open
forever.
The number of connection a database can handle is limited and
can't be increased without a restart.
Pooling is enabled by default because it increase the performance by not
having to reopen a connection for each query, but in production we
already use an external connection pooler (pg_pool) that allow us to
share connection between multiple process or even box.

With this change we will be able to increased the number of jormungandr
processes without impacting the database too much