Jormungandr: move schedule update methods in the proxy
Allow us to have a custom merging between the datas from kraken and the
datas from the proxy. If we just want the realtime proxy datas to be
active for one day or even one hour, the proxy can define which datas
from kraken it is removing and define how it does the merging.
An example is in proxy_realtime_tests.