Kraken: add unit test on next_stop_time / schedules
Add some unit test failing with long duration on stop_schedules /
route_schedules and next_stop_time. If there is a gap of more than one
day to the next service kraken will stop returning new datetime even if
there is more after. It's because right now we are only looking for the
next departure on today and tomorrow. See issue #2279.