Kraken: expand next departure search with bound
This change the way (next/prev)_valid_(discrete/freq) look for a
departure or arrival. We are currently only looking today and the day
after / before. Now we take advantage of the existing bound to expand
the horizon if necessary. We look on the current day and loop on each
day until the bound is reached.
If no bound is passed we keep the current implied bound of tomorrow at
midnight. We make sure to not cross the end production date.