kraken: ignore envvar not defined for kraken
Kubernetes define a lot of environement variable for each service.
For example if a service "kraken-default" exist the following envvar
(it may not be the exact naming, but you get the idea) will exist inside
each pod:
  - KRAKEN_DEFAULT_SERVICE_PORT
  - KRAKEN_DEFAULT_SERVICE_HOST

Obviously these variable aren't used by kraken, but boost program
options will consider it as an error if they are present.
We still have to not define service named "kraken-general",
"kraken-broker" or "kraken-chaos", but it seem pretty unlikely that we
will do that.