Prometheus integration
This will not build as I didn't made the submodule and the cmake stuff.
It use this lib: https://github.com/pingcap/rust-prometheus

Their is a few things that I don't like in what I did for now
The creation of the histogram with the `Add` method will acquire a lock,
so I don't like the idea of using it inside the worker loop but maybe
that I am a bit paranoid about it. But at the same time having to
declare every metric in the class `Metrics` is cumbersome and will force
use to have a big if: not pretty. Last possibility, a readonly map so it
can use between threads that is handled by the `Metrics` class, but it
also require to known every label that will be use at compile time.

Also I'm not that confident about that lib, so maybe having a way to
disable it will be useful, only configuration or IFDEF to be safe?