fix failing ridehsharing test with protobuf 2.6
The behavior between protobuf 3 and protobuf 2.6 isn't the same when
setting an integer into a float field.
In protobuf 2.6 it will stay an int
and as such we loose the decimal in the output.
In protobuf 3, it is cast to a float and we have the decimal in the
output.
So, long story short, let use float in the test and not be bothered we
this.

Fix: #2360