Jormun: slight modification to timeo to make it easy for gomeo
Timeo use a non standard way of formatting parameters: `StopDescription`
is an "object" that contains params between the `?` and `;`. As you can
guess there is not a single implementation of something like this.
Standard implementation will consider that the value of
`StopDescription` end after the next `&`, so we "loose" the first
params.
This PR put `StopTimeType` in first position since it's the less useful
parameter of all, since for us it is obvious that we call timeo to get
realtime data. Doing so make it very easy for [goemo](https://github.com/kinnou02/gomeo)
to parse the query.