kraken: bind zmq socket after the load of data
Before this change the zmq socket was bind at the startup of kraken,
before loading the data, this mean that it will start accepting request
that cannot yet be served.
With this change the socket is bind at the last possible moment, so
every request accepted by the server will be processed.