Ed: match stop_area and poi by using admin boundary if possible
When OSM data is used we git the boundary of each admins, in that case
we want to use these boundary to associate stop_points to an admin.

This doesn't use any kind of spacial indexes, so it might be slower than
before.

If we do not find any admin by using the boundary we still keep the old
approach as a fallback, so if there is no boundaries it work as before.