jormun: Round second of datetime before calling realtime proxy to improve cachability
The cache key used with realtime proxy contains the url called, so it
does contains the datetime with the seconds, it make cache only working
for one second, that's not very good...
This PR round the second to a multiple of 30, giving us 30 seconds of
cache, this is also configurable with the `step` parameter.

We may want to rename this params as it's not very clear what it do
based on it's name, so if you have an idea I take it.