raptor: increase number of try in raptor for journeys without transfers
When requesting journeys with min_nb_journeys or timeframe_duration we
loop in kraken to find most of the journeys in one call. To prevent
infinite loop we limit to 100 iterations. This PR increase the maximum
of iteration when the request doesn't allow any transfers
(max_transfers=0) as this type of journeys is very fast to compute and
is mostly used to build time schedules between two stops.

This PR also add a lot of unit test for the keep_going function