Jormun: remove an useless log
Jormungandr log a warning if it doesn't find a coverage in it's database
and use the default value defined in the code. This log was also written
when the requested value was `None`.
This PR isolate the test for the log and only write it if the instance
isn't in the database: this should never happen in production.