Jormun: add a memory cache layer and migration to flask_caching
Flask_cache is unmaintained since 2014, it has been forked and maintained has flask_caching.
This PR also add a new cache instance named `memory_cache` that aims to
provide a first layer of cache that is faster (but local) than the
traditional `cache` that is (most of the time) backed by a redis.

The `memory_cache` is only used to reduce load on the database and as
such isn't used for realtime proxies. This is mostly due to the fact
that the priviligied backend for this cache will be the [uwsgi
cache](https://uwsgi-docs.readthedocs.io/en/latest/Caching.html) that
require to be configured with a maximum number of item and doesn't
provide both LRU and time based expiration at the same time.