Sytral realtime based on custom ws
Realtime departures for sytral, this used sytral-rt: https://github.com/CanalTP/sytralrt
This connector is based on the cleverage one as we got mostly the same
information. It has the same drawback of not being able to distinguish
between the departure of two route of the same line at the same
stop_point.