Kraken: prioritize main stop_area over the others stop_area
With the change we reintroduce one of the rules used by navitia 1.
It allow data adminnistrators to increase the weight of some stop_area
by setting them as main stop_area.
This is a hack to make JLP happy: we need to found a better way to
implement this as main stop_area are used for admin to admin journeys
and geocoding.
You may want to set a coach station as the departure/arrival of a city
when doing long distance journeys, but it's a "minor" stop in the urban
network and even in the national network