time_tables: add new tests on loop, terminuses
Following issue 2289 when we have a line passing multiple times at a
stop_point we don't have any date_times on it since the check is pretty
simple, just a comparison of the stop_point with the last one of the jp.
The goal is to be able to ouput departures on those stops.
This cause some other changes on the different ResponseStatus which we
test too.