Jormungandr: implement transient zmq socket type like http keepalive
This PR improve the transient zmq socket type by making them less
transient but not definitive like the `persistent` socket type.

It look a bit like http keepalive: opened sockets are reused if
possible, after some time idle the socket will be closed.
The closing of the socket is done by a dedicated greenlet that is run
every `ZMQ_SOCKET_REAPER_INTERVAL` seconds (default: 10) and it will
close the socket if it hasn't been used since `ZMQ_SOCKET_TTL_SECONDS`
seconds (default: 10).
This will help to distribute the load on every kraken instances while
reducing the connection opened that aren't used.

Every instance has a *socket reaper thread* (greenlet in fact) that
handle the socket of this instance. This greenlet use gevent so it
require to have gevent enabled on the uwsgi level, more or less like
distributed, so it's not so bad.
Without gevent at the uwsgi level the greenlet will only be scheduled
when the jormungandr process is waiting on a response from kraken.

I also upgraded gevent as the last version is retrocompatible, faster
and with less bug.