kraken: refactoring of autocomplete to improve performance
Previously the scoring part of the algo was done on the protobuf, so the
response has to be build before the selection of the results. So for 30
requested results we would potentially serialize 1200 items before
scoring it. When serializing a Way we have to project the adress along
it's way: obviously this take a lot of time.

With this PR the scoring is only done with `fl_quality` and only the
returned results are serialized.
It also implements a few others optimizations, the bigest is only
available on search_type==0 (prefix mode): we stop the search once we
have enough items to return. This is possible since each type is
associated to an order and that we never return a object of a lower
order if there is enough objects with an higher order.