Jormun: log request-id provided in the request
If a client provide a request-id with the `x-request-id` header this one
will be logged. This will help to trace request between systems.
This external request id doesn't change our request id that is generated
by jormungandr at the start of a request as their is no easy way to
ensure that the provided id is unique.

This is only logged if json log is activated.