jormun: disable uwsgi timer waiting for proper fix
uwsgi has a undocumented limit on the number of timer than can be
created in a uwsgi "context", set to 64, after that creating the timer
will raise an exception that terminate all uwsgi processes.
The current implementation of transient socket create one timer per
uwsgi process.
We learned the hard way that this limit exist when our uwsgi in
production started to die at startup...

This doesn't fix the problems, it only disable uwsgi timers to let me
some time for finding a real solution.

This will be hotfixed in 2.68.2