jormun: refactoring of zmq socket reaper
This PR aim to fix the problem with the current implementaiton of the
zmq socket reaper: it doesn't work in production...

There is a hard limit on the number of uwsgi timer that can be
instanced in uwsgi: 64, when trying to create more uwsgi will raise a
ValueError that will kill jormungandr.
Also uwsgi timer aren't perfect they are dispatched on workers randomly
and not by using round robbin, so it's possible that a worker never get
the chance to handle a signal while another will get it each time. (This
comportment can also be observed with request handling)

This PR make multiple change to improve the previous implementation:

  1) Always start a gevent greenlet that should close every reaper interval
     to close idle sockets.

  2) If we are running in uwsgi without the gevent event loop enable
     create an uwsgi signal handler to close the sockets that will be
     triggered by an uwsgi timer

  3) If we can't create an uwsgi timer it's not that bad, so jormun
     continue to run in that case.

Uwsgi timers are a failsafe: By default uwsgi doesn't use the gevent
event loop, so an idle worker can't run the socket reaper greenlet
periodically, it's only run when waiting for an external service like
kraken.
At most 64 timers will be created it there is more processes they can
still handle a signal, so at most 64 processes will close their socket
every reaper_interval.