jormun: stop reporting bragi error as newrelic exception
When bragi is a bit slow there is a lot of timeout on the reverse done
for /journeys, currently these timeout are reported to newrelic as
exceptions even if they do not cause the response to be on error.
This behavior is a bit misleading for operation people that see 20-30%
of error in newrelic while in fact it's just that the reverse is
failing: it's not that bad.

This PR create a new type of events in newrelic to track
`GeocodeJsonUnavailable` errors (timeouts, 503, etc...), other kind
of errors (`GeocodeJsonError`) are still reported as exception.

The newrelic event is `autocomplete_status`, is has the following
properties:
  - status: ok/unavailable/error
  - type: geocodejson or kraken
  - cause: exception message in case of status being unavailable or error