Jormun: Fix count on /departures with realtime proxies
This is a follow up on https://github.com/CanalTP/navitia/pull/2645
The count parameter wasn't used correctly on departures when realtime
proxies were involved.
This add a post action that will enforce the count, so we never get more
than `count` departure.

We could change the realtime proxies to let them know about the
differences between /departures and /stop_schedules to optimize calls
made to external services. It's a lot of change that I won't make for
now.