Jormun: enable keepalive on http call to bragi
This aims to improve performance by reducing the number of tcp
connection to open.
I have done some testing from the office network to the production:
without keepalive response time for a simple autocomplete is around
95ms. With keepalive it's around 80ms.
There is obvisouly some overhead in this test from our vpn and internet,
on production there will be way less hop.

Requests handle everything once a session is enabled, it will reopen the
connection if needed, etc.

In case of a problem it's possible to disable the keepalive in the
configuration by setting the parameter "disable_keepalive" to true in
the geocodejson convfiguration