jormun: restrict bragi autocompletion to the coverage
When doing a /places on a coverage if no shape as been provided by the
user we use the shape of the coverage to prevent showing objects that
are outside of the coverage.
The shape of the coverage is simplified to speedup the search.
An evolution has to be made on bragi to not apply the shape's filter on
stop areas.