kraken: add failing disruption tests
There is two cases not currently handled by navitia for disruptions.
If we have a trip with two (or more) consecutive stop_points in the same stop_area we
are only impacting the first / last stop_point in a line section if this
stop_area is the start / end of the section.
If we have a lollipop VJ with the same stop_point present multiple
time, if one of the occurrence of the stop_point is impacted by a
section, all occurrences of the stop_point will be impacted.